split ExecuteStep and ApproveStep
