+++ components/Tweet/Proposals/index.jsx
@@ -6,7 +6,7 @@ import { EducationTitle } from 'common-util/Education/EducationTitle';
 import { Proposal } from './Proposal';
 import { useCentaursFunctionalities } from '../../CoOrdinate/Centaur/hooks';
 
-const Proposals = () => {
+export const Proposals = () => {
   const { isMemoryDetailsLoading: isLoading, currentMemoryDetails } = useCentaursFunctionalities();
   const proposals = currentMemoryDetails?.plugins_data?.scheduled_tweet?.tweets || [];
   const sortedProposals = proposals.sort(
@@ -45,5 +45,3 @@ const Proposals = () => {
     </>
   );
 };
-
-export default Proposals;
